Move ViewModel to the general interface_adapter package
